requirements.txt
fastapi==0.120.2
uvicorn==0.38.0

python-socketio==5.14.3
python-engineio==4.12.3

pydantic==2.12.3
pydantic_core==2.41.4

loguru==0.7.3
PyJWT==2.10.1

requests==2.32.5
websocket-client==1.9.0

pytest==8.4.2
pytest-asyncio==1.2.0
